Added barrelling for play button
